refactor: MVC로직분리
